requirements.txt
aldryn-apphooks-config==0.6.0
Django==3.0
django-appdata==0.3.0
django-classy-tags==2.0.0
django-cms==3.7.4
django-filer==1.7.1
django-formtools==2.2
django-js-asset==1.2.2
django-meta==1.7.0
django-mptt==0.11.0
django-parler==2.1
django-polymorphic==2.1.2
django-sekizai==2.0.0
django-sortedm2m==3.0.2
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==1.2.0
djangocms-blog==1.1.1
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
easy-thumbnails==2.7
html5lib==1.1
lxml==4.5.2
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
webencodings==0.5.1
drf-yasg==1.20.0
